+++ src/renderer/helpers/Normalizers.js
